[Refactor] Office 엔티티 도메인 레벨로 변경
