+++ src/main/java/com/final_10aeat/domain/admin/entity/Admin.java
@@ -1,6 +1,7 @@
 package com.final_10aeat.domain.admin.entity;
 
 import com.final_10aeat.common.enumclass.MemberRole;
+import com.final_10aeat.domain.office.entity.Office;
 import com.final_10aeat.global.entity.SoftDeletableBaseTimeEntity;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
