+++ src/main/java/com/final_10aeat/domain/admin/service/AdminService.java
@@ -2,10 +2,10 @@ package com.final_10aeat.domain.admin.service;
 
 import com.final_10aeat.domain.admin.dto.request.CreateAdminRequestDto;
 import com.final_10aeat.domain.admin.entity.Admin;
-import com.final_10aeat.domain.admin.entity.Office;
-import com.final_10aeat.domain.admin.exception.OfficeNotFoundException;
+import com.final_10aeat.domain.office.entity.Office;
+import com.final_10aeat.domain.office.exception.OfficeNotFoundException;
 import com.final_10aeat.domain.admin.repository.AdminRepository;
-import com.final_10aeat.domain.admin.repository.OfficeRepository;
+import com.final_10aeat.domain.office.repository.OfficeRepository;
 import com.final_10aeat.domain.member.dto.request.MemberLoginRequestDto;
 import com.final_10aeat.common.enumclass.MemberRole;
 import com.final_10aeat.domain.member.exception.EmailDuplicatedException;
