+++ src/main/resources/logback-spring.xml
@@ -33,7 +33,7 @@
 
   <property name="LOGS" value="./logs"/>
 
-  <root level="DEBUG">
+  <root level="INFO">
     <appender-ref ref="CONSOLE"/>
     <appender-ref ref="RollingFile"/>
   </root>
