+++ src/test/java/com/final_10aeat/domain/admin/docs/AdminControllerDocsTest.java
@@ -18,9 +18,9 @@ import com.final_10aeat.docs.RestDocsSupport;
 import com.final_10aeat.domain.admin.controller.AdminController;
 import com.final_10aeat.domain.admin.dto.request.CreateAdminRequestDto;
 import com.final_10aeat.domain.admin.entity.Admin;
-import com.final_10aeat.domain.admin.entity.Office;
+import com.final_10aeat.domain.office.entity.Office;
 import com.final_10aeat.domain.admin.repository.AdminRepository;
-import com.final_10aeat.domain.admin.repository.OfficeRepository;
+import com.final_10aeat.domain.office.repository.OfficeRepository;
 import com.final_10aeat.domain.admin.service.AdminService;
 import com.final_10aeat.domain.member.dto.request.MemberLoginRequestDto;
 import com.final_10aeat.global.security.jwt.JwtTokenGenerator;
